Commit para poder sacar las respuestas a una pregunta de BD

+++ hibernate-lib/src/main/java/kadamm/hibernate/dao/PreguntesDao.java
@@ -63,7 +63,7 @@ public class PreguntesDao {
     }
 
     // Create method to get preguntes from database with hibernate
-    public Preguntes getPreguntaById(int id) {
+    public Preguntes getPreguntaById(long id) {
         Session session = HibernateUtil.getSessionFactory().openSession();
         Preguntes preguntes = null;
         try {
